+++ src/django_components/component_media.py
@@ -5,6 +5,7 @@ from copy import copy
 from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Callable, Dict, List, Literal, Optional, Protocol, Tuple, Type, Union, cast
+from weakref import WeakKeyDictionary
 
 from django.contrib.staticfiles import finders
 from django.core.exceptions import ImproperlyConfigured
@@ -398,7 +399,11 @@ def _get_comp_cls_attr(comp_cls: Type["Component"], attr: str) -> Any:
     return None
 
 
-media_cache: Dict[Type["Component"], MediaCls] = {}
+# NOTE: We use weakref to avoid issues with lingering references.
+if sys.version_info >= (3, 9):
+    media_cache: WeakKeyDictionary[Type["Component"], MediaCls] = WeakKeyDictionary()
+else:
+    media_cache: WeakKeyDictionary = WeakKeyDictionary()
 
 
 def _get_comp_cls_media(comp_cls: Type["Component"]) -> Any: